pyproject.toml
[project]
name = "json-spike"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "lxml>=5.3.1",
    "pyodbc>=5.2.0",
    "python-dotenv>=1.0.1",
    "sqlalchemy",
]
